start_code/doc_maker/requirements.txt
openai>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
click>=8.1.7
python-dotenv>=1.0.0
loguru>=0.7.0
